pyproject.toml
[tool.poetry]
name = "fastapi_template"
version = "0.1.0"
description = "Template for fastapi"
authors = ["DavianYang <kothantyarzarhein@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.14.0"
fastapi = "^0.65.2"
pydantic = "^1.8.2"
PyJWT = "^2.1.0"
databases = "^0.4.3"
asyncpg = "^0.23.0"
psycopg2-binary = "^2.8.6"
alembic = "^1.6.5"
docker = "^5.0.0"
asgi-lifespan = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
autoflake = "^1.4"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
docker = "^5.0.0"
httpx = "^0.18.1"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=100
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
